Pagination
Why:
So API consumer can have a choice of number of bucket lists to see on a particular page.
So that the page will not be too crowded, or too long.

This issue is fixed by:
Making use of Active records limit and offset for database queries.
Inclusion of params page and limits where users can specify the number they want
Setting of default number of items to display and the page to display(defaults page:1 limit:20)

[Finshes #108494468]